+++ app/Models/Settings/Users/Permission.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Settings\Users;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Models\User;
 
 class Permission extends Model
 {
